Initial Commit
Develop a sample app to demonstrate your expertise in test-driven
development, object oriented design, and the iOS platform. 

Please submit a .zip file of the xCode workspace containing all tests
and code.

The code must be written in Objective-C.

The app should allow the user to choose a city and state and view the 3
day weather forecast. The forecast for each day should display the high
temp, low temp, and textual description.

Use this web service that is free for developers .

http://www.wunderground.com/weather/api/?ref=twc